close issue#3
